+++ domain/Form/Actions/CreateFormAction.php
@@ -20,6 +20,7 @@ class CreateFormAction
             'blueprint_id' => $formData->blueprint_id,
             'name' => $formData->name,
             'store_submission' => $formData->store_submission,
+            'uses_captcha' => $formData->uses_captcha,
         ]);
 
         foreach ($formData->form_email_notifications ?? [] as $formEmailNotification) {