+++ domain/Role/Models/Role.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Domain\Role\Models;
 
-use AlexJustesen\FilamentSpatieLaravelActivitylog\Contracts\IsActivitySubject;
 use Domain\Support\ConstraintsRelationships\Attributes\OnDeleteCascade;
 use Domain\Support\ConstraintsRelationships\Attributes\OnDeleteRestrict;
 use Domain\Support\ConstraintsRelationships\ConstraintsRelationships;
@@ -41,7 +40,7 @@ use Spatie\Activitylog\Traits\LogsActivity;
     OnDeleteCascade(['permissions']),
     OnDeleteRestrict(['users'])
 ]
-class Role extends \Spatie\Permission\Models\Role implements IsActivitySubject
+class Role extends \Spatie\Permission\Models\Role
 {
     use LogsActivity;
     use ConstraintsRelationships;
@@ -52,10 +51,5 @@ class Role extends \Spatie\Permission\Models\Role implements IsActivitySubject
             ->logUnguarded()
             ->logOnlyDirty()
             ->dontSubmitEmptyLogs();
-    }
-
-    public function getActivitySubjectDescription(Activity $activity): string
-    {
-        return 'Role: '.$this->name;
     }
 }